switched to inference
